Move towards NamespacedProvider setup.
Providers reference Secrets and ConfigMaps. These are namespace bound,
so we need to make the provider namespace bound as well.

In a first phase, we'll default to the namespace where the Monitor is
registered (if it's automatically generated). We'll then move towards a
system where you can specify the namespace, which will allow for
reusability.