Add fix for StatusCake TestID.
We were seeing issues where on data changes, the Operator would create a
new monitor. This was caused because StatusCake returns no ID when there
are changes to the object. This translates into a 0 within the client.

We used to return this 0 id, which then lead to a non existing error,
causing us to create a new monitor with StatusCake.